requirements.txt
langchain==0.3.12
langchain_community==0.3.12
langchain-openai==0.2.12
pdfplumber==0.11.5
pandas==2.2.3
python-pptx==1.0.2
python-docx==1.1.2
olefile==0.47
tabulate==0.9.0
opensearch-py==2.8.0
sentence-transformers==3.3.1
ollama==0.4.5
redis==5.2.1
jupyter==1.1.1
notebook==7.3.2
#huggingface-hub==0.27.1
#datasets==3.2.0
